Rename CityProvider to CityService
