+++ AspireProjects/AppHost/AppHost.csproj
@@ -14,7 +14,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\..\src\CityProvider\CityProvider.csproj" />
+    <ProjectReference Include="..\..\src\CityService\CityService.csproj" />
     <ProjectReference Include="..\..\src\TemperatureService\TemperatureService.csproj" />
     <ProjectReference Include="..\..\src\WeatherForcast\WeatherForcast.csproj" />
   </ItemGroup>
