+++ AspireProjects/AppHost/Program.cs
@@ -2,12 +2,13 @@
 
 var builder = DistributedApplication.CreateBuilder(args);
 
-var cityProvider = builder.AddProject<CityProvider>("cityprovider");
+var cityservice = builder.AddProject<CityService>("cityservice");
 var temperatureProvider = builder.AddProject<TemperatureService>("temperatureservice");
 
 builder
     .AddProject<WeatherForcast>("weatherforcast")
-    .WithReference(cityProvider)
+    .WithReference(cityservice)
     .WithReference(temperatureProvider);
+
 var app = builder.Build();
 app.Run();
