+++ PactDemo.sln
@@ -17,8 +17,6 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "WeatherForcast", "src\WeatherForcast\WeatherForcast.csproj", "{C8AD8F0F-06AD-4775-8BDC-027714B6EAA1}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CityProvider", "src\CityProvider\CityProvider.csproj", "{D3DAB8E1-2017-4E7B-8E54-FEEACF497974}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "WeatherForcastContractTests", "test\WeatherForcastContractTests\WeatherForcastContractTests.csproj", "{8823BF0B-D9C3-4359-B9A6-BA620D6EEB28}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "AspireProjects", "AspireProjects", "{6521E5A4-D661-43C7-A837-BAF2BC7048BA}"
@@ -37,6 +35,8 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DemoConfigurations", "src\S
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TemperatureService", "src\TemperatureService\TemperatureService.csproj", "{B59B941A-69DA-4660-BF8A-C9B61B80B35C}"
 EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CityService", "src\CityService\CityService.csproj", "{8A5B9DAD-DC4A-4E16-AB44-71D67AE6C776}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -47,10 +47,6 @@ Global
 		{C8AD8F0F-06AD-4775-8BDC-027714B6EAA1}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{C8AD8F0F-06AD-4775-8BDC-027714B6EAA1}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{C8AD8F0F-06AD-4775-8BDC-027714B6EAA1}.Release|Any CPU.Build.0 = Release|Any CPU
-		{D3DAB8E1-2017-4E7B-8E54-FEEACF497974}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{D3DAB8E1-2017-4E7B-8E54-FEEACF497974}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{D3DAB8E1-2017-4E7B-8E54-FEEACF497974}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{D3DAB8E1-2017-4E7B-8E54-FEEACF497974}.Release|Any CPU.Build.0 = Release|Any CPU
 		{8823BF0B-D9C3-4359-B9A6-BA620D6EEB28}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{8823BF0B-D9C3-4359-B9A6-BA620D6EEB28}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{8823BF0B-D9C3-4359-B9A6-BA620D6EEB28}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -79,13 +75,16 @@ Global
 		{B59B941A-69DA-4660-BF8A-C9B61B80B35C}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{B59B941A-69DA-4660-BF8A-C9B61B80B35C}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{B59B941A-69DA-4660-BF8A-C9B61B80B35C}.Release|Any CPU.Build.0 = Release|Any CPU
+		{8A5B9DAD-DC4A-4E16-AB44-71D67AE6C776}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{8A5B9DAD-DC4A-4E16-AB44-71D67AE6C776}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{8A5B9DAD-DC4A-4E16-AB44-71D67AE6C776}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{8A5B9DAD-DC4A-4E16-AB44-71D67AE6C776}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
 		{C8AD8F0F-06AD-4775-8BDC-027714B6EAA1} = {575244E4-337F-45C4-886E-0CC27F8F892A}
-		{D3DAB8E1-2017-4E7B-8E54-FEEACF497974} = {575244E4-337F-45C4-886E-0CC27F8F892A}
 		{8823BF0B-D9C3-4359-B9A6-BA620D6EEB28} = {4D21B2A1-8B90-48D3-9D49-7C48B81AE481}
 		{F437E33F-5FD8-42C1-8DE9-59591D2C077F} = {6521E5A4-D661-43C7-A837-BAF2BC7048BA}
 		{FD855778-252A-48F7-998E-D8B2B91004C0} = {6521E5A4-D661-43C7-A837-BAF2BC7048BA}
@@ -94,6 +93,7 @@ Global
 		{B7D87FB5-427D-4FE4-94A6-E35B5FEA1877} = {575244E4-337F-45C4-886E-0CC27F8F892A}
 		{EE36A974-48E9-47B1-9BA0-F9C7C74ACC62} = {B7D87FB5-427D-4FE4-94A6-E35B5FEA1877}
 		{B59B941A-69DA-4660-BF8A-C9B61B80B35C} = {575244E4-337F-45C4-886E-0CC27F8F892A}
+		{8A5B9DAD-DC4A-4E16-AB44-71D67AE6C776} = {575244E4-337F-45C4-886E-0CC27F8F892A}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {7BDAD170-243C-4FCA-95DE-59B8960AD6A8}
