--- src/CityProvider/CityProvider.http
@@ -1,6 +1,0 @@
-@CityProvider_HostAddress = http://localhost:5186
-
-GET {{CityProvider_HostAddress}}/weatherforecast/
-Accept: application/json
-
-###
