--- src/CityProvider/CityProviderMarker.cs
@@ -1,3 +1,0 @@
-﻿namespace CityProvider;
-
-public sealed class CityProviderMarker { }
