+++ src/CityService/CityService.http
@@ -1,0 +1,6 @@
+@CityService_HostAddress = http://localhost:5186
+
+GET {{CityService_HostAddress}}/weatherforecast/
+Accept: application/json
+
+###
