+++ src/CityService/CityServicerMarker.cs
@@ -1,0 +1,3 @@
+﻿namespace CityService;
+
+public sealed class CityServiceMarker { }
