+++ src/CityService/Program.cs
@@ -1,4 +1,4 @@
-﻿using CityProvider;
+﻿using CityService;
 
 var app = Startup.WebApp(args);
 app.Run();
