+++ src/CityService/V1/Cities/GetCitiesResponse.cs
@@ -1,5 +1,5 @@
 ﻿using Clients.V1.Models;
 
-namespace CityProvider.V1.Cities;
+namespace CityService.V1.Cities;
 
 public sealed record GetCitiesResponse(ICollection<City> Cities) { }
