+++ test/CityProviderContractTests/ServiceCollectionExtensions.cs
@@ -4,7 +4,7 @@ using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
-using CityProviderStartup = CityProvider.Startup;
+using CityProviderStartup = CityService.Startup;
 
 namespace CityProviderContractTests;
